refactor: calculate interval for overtime requests
https://gravitee.atlassian.net/browse/APIM-7354
