+++ gravitee-apim-rest-api/gravitee-apim-rest-api-portal/gravitee-apim-rest-api-portal-rest/src/main/resources/portal-openapi.yaml
@@ -10,7 +10,7 @@ info:
     license:
         name: Apache 2.0
         url: http://www.apache.org/licenses/LICENSE-2.0
-    version: "4.6.0-archi-401-opentelemetry-SNAPSHOT"
+    version: "4.6.0-SNAPSHOT"
 servers:
     - url: http://localhost:8083/portal/environments/{envId}
       description: The portal API for a given environment
