+++ gravitee-apim-rest-api/gravitee-apim-rest-api-service/src/main/java/io/gravitee/apim/core/analytics/use_case/SearchResponseTimeUseCase.java
@@ -22,14 +22,13 @@ import io.gravitee.apim.core.api.exception.ApiInvalidDefinitionVersionException;
 import io.gravitee.apim.core.api.exception.ApiNotFoundException;
 import io.gravitee.apim.core.api.exception.TcpProxyNotSupportedException;
 import io.gravitee.apim.core.api.model.Api;
+import io.gravitee.apim.core.utils.DurationUtils;
 import io.gravitee.definition.model.DefinitionVersion;
 import io.gravitee.rest.api.service.common.ExecutionContext;
 import io.reactivex.rxjava3.core.Maybe;
 import io.reactivex.rxjava3.core.Single;
 import java.time.Duration;
 import java.time.Instant;
-import java.util.Collection;
-import java.util.Comparator;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 
@@ -40,31 +39,10 @@ public class SearchResponseTimeUseCase {
     private final AnalyticsQueryService analyticsQueryService;
     private final ApiCrudService apiCrudService;
 
-    private static final Collection<IntervalData> INTERVAL = List.of(
-        new IntervalData(Duration.ofMinutes(5), Duration.ofSeconds(10)),
-        new IntervalData(Duration.ofMinutes(30), Duration.ofSeconds(15)),
-        new IntervalData(Duration.ofHours(1), Duration.ofSeconds(30)),
-        new IntervalData(Duration.ofHours(3), Duration.ofMinutes(1)),
-        new IntervalData(Duration.ofHours(6), Duration.ofMinutes(2)),
-        new IntervalData(Duration.ofHours(12), Duration.ofMinutes(5)),
-        new IntervalData(Duration.ofDays(1), Duration.ofMinutes(10)),
-        new IntervalData(Duration.ofDays(3), Duration.ofMinutes(30)),
-        new IntervalData(Duration.ofDays(7), Duration.ofHours(1)),
-        new IntervalData(Duration.ofDays(14), Duration.ofHours(3)),
-        new IntervalData(Duration.ofDays(30), Duration.ofHours(6)),
-        new IntervalData(Duration.ofDays(90), Duration.ofHours(12))
-    );
-
     public Single<Output> execute(ExecutionContext executionContext, Input input) {
         Instant to = input.to();
         Instant from = input.from();
-        Duration duration = Duration.between(from, to);
-        Duration interval = INTERVAL
-            .stream()
-            .filter(id -> id.dataDuration().compareTo(duration) <= 0)
-            .max(Comparator.comparing(IntervalData::dataDuration))
-            .map(IntervalData::interval)
-            .orElse(Duration.ofDays(1));
+        Duration interval = DurationUtils.buildIntervalFromTimePeriod(from, to);
 
         return validateApiRequirements(input)
             .flatMapMaybe(api -> analyticsQueryService.searchAvgResponseTimeOverTime(executionContext, api.getId(), from, to, interval))
@@ -98,6 +76,4 @@ public class SearchResponseTimeUseCase {
     public record Input(String apiId, String environmentId, Instant from, Instant to) {}
 
     public record Output(Instant from, Instant to, Duration interval, List<Long> data) {}
-
-    private record IntervalData(Duration dataDuration, Duration interval) {}
 }
